fix(test): convert wait-on to task and improve tests configuration (#2135)
