+++ centreon/tests/e2e/cypress/tsconfig.json
@@ -4,7 +4,7 @@
     "baseUrl": "..",
     "esModuleInterop": true,
     "target": "es2020",
-    "types": ["cypress", "node"]
+    "types": ["cypress", "cypress-wait-until", "node"]
   },
   "include": [
     "**/*.ts"
