+++ centreon/tests/e2e/package.json
@@ -12,21 +12,15 @@
     "eslint:fix": "pnpm eslint --fix"
   },
   "devDependencies": {
-    "@badeball/cypress-cucumber-preprocessor": "^18.0.4",
-    "@cypress/webpack-preprocessor": "^5.17.1",
-    "@swc/core": "^1.3.78",
-    "@types/cypress-cucumber-preprocessor": "^4.0.1",
-    "@types/dockerode": "^3.3.19",
-    "@types/node": "^20.5.3",
-    "cypress": "^12.17.4",
+    "@badeball/cypress-cucumber-preprocessor": "^18.0.5",
+    "@types/cypress-cucumber-preprocessor": "^4.0.2",
+    "@types/node": "^20.5.9",
+    "cypress": "^13.1.0",
     "cypress-wait-until": "^2.0.1",
-    "dockerode": "^3.3.5",
     "mochawesome": "^7.1.3",
     "path": "^0.12.7",
     "shell-exec": "^1.1.2",
-    "swc-loader": "^0.2.3",
-    "typescript": "^5.1.6",
-    "webpack": "^5.88.2"
+    "typescript": "^5.2.2"
   },
   "cypress-cucumber-preprocessor": {
     "nonGlobalStepDefinitions": true
